moved set(node) from Subject to Page
